transfered files to components folder
